auth_services/auth_service/requirements.txt
# Django Framework
Django==4.2

# Django REST Framework
djangorestframework==3.14.0
drf-yasg==1.21.7

# Authentication (JWT)
djangorestframework-simplejwt==5.3.1

# PostgreSQL Driver
psycopg2-binary==2.9.9

# ImageField Support
Pillow==10.2.0

# CORS (if using frontend)
django-cors-headers==4.3.0

# Social Auth (Google OAuth)
social-auth-app-django==5.4.1

# Environment Variables
python-dotenv==1.0.0

# Email, Requests
requests==2.31.0

# For RabbitMQ (future usage)
pika==1.3.2

# OTP / MFA utilities
pyotp==2.9.0
qrcode==7.4.2

# For async
asgiref==3.7.2
